Placement/requirement.txt
numpy
torch
torchvision
pillow
scikit-image
fire
matplotlib
opencv-python
numba
easydict
cython
tqdm
pyquaternion
pandas